requirements.txt
# Core RAG components
langchain==0.1.0
langchain-community==0.0.10

# Vector store
chromadb==0.4.22

# Document loaders
pypdf==3.17.4
python-docx==1.1.0
unstructured==0.11.8

# FREE LLM provider - Ollama
ollama==0.1.7

# FREE Embeddings - HuggingFace
sentence-transformers==2.3.1
huggingface-hub==0.20.3

# Web framework
streamlit==1.29.0
streamlit-chat==0.1.1

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1

# Development
pytest==7.4.3
black==23.12.1

# Additional utilities
numpy==1.26.3